+++ src/window_children_layout.h
@@ -1,0 +1,33 @@
+/**
+ * File:   window_children_layout.h
+ * Author: AWTK Develop Team
+ * Brief:  window_children_layout source
+ *
+ * Copyright (c) 2018 - 2021  Guangzhou ZHIYUAN Electronics Co.,Ltd.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * License file for more details.
+ *
+ */
+
+/**
+ * History:
+ * ================================================================
+ * 2021-05-16 Li XianJing <xianjimli@hotmail.com> created
+ *
+ */
+
+#ifndef WINDOW_CHILDREN_LAYOUT_H
+#define WINDOW_CHILDREN_LAYOUT_H
+
+#include "awtk.h"
+
+BEGIN_C_DECLS
+
+ret_t window_children_layout_open(void);
+
+END_C_DECLS
+
+#endif /*WINDOW_CHILDREN_LAYOUT_H*/